VCV-219: separate info panel and info panel row into own components
